+++ frontend/package.json
@@ -177,6 +177,7 @@
         "lodash.merge": "^4.6.2",
         "lodash.sortby": "^4.7.0",
         "lodash.uniqby": "^4.7.0",
+        "lodash.throttle": "^4.1.1",
         "maplibre-gl": "^3.5.1",
         "marked": "^15.0.11",
         "mathjax-full": "3.2.2",
@@ -251,6 +252,7 @@
         "@types/lodash.merge": "^4.6.9",
         "@types/lodash.sortby": "^4.7.9",
         "@types/lodash.uniqby": "^4.7.9",
+        "@types/lodash.throttle": "^4.1.1",
         "@types/md5": "^2.3.0",
         "@types/node": "^22.13.14",
         "@types/papaparse": "^5.3.8",
